backend/requirements.txt
# SpecForge AI - Python Dependencies
# ====================================

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
greenlet>=3.0.0

# AI Agent Framework
crewai>=0.121.0

# LLM Providers
openai>=1.60.0
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# SSE Support
sse-starlette>=2.2.0

# File Operations
aiofiles>=24.1.0
python-multipart>=0.0.18
